backend/requirements.txt
# Web框架
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
gunicorn==21.2.0

# 数据库
pymysql==1.1.0
redis==5.0.1
DBUtils==3.1.0

# 数据采集
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
fake-useragent==1.4.0

# 定时任务
APScheduler==3.10.4

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 七牛云存储
qiniu==7.12.1

# 工具
python-dotenv==1.0.0
loguru==0.7.2
pyjwt==2.8.0
cryptography==41.0.7
